Updated the DepthByColor function path
The Depth by color method actually kind of works. We now get the shape outputted in the blender scene and all the points have different Z values mostly as intended.

Updated serval UI elements in base UI to make this version seem closer to the intended Figma representation.

Moved the User data into the Database UI because it was in the UI operations where it wasn't being used.

Moved the DatabaseUIMenu into the base UI where it should be and I have the function grab the user data.

Removed the Access DB Custom Panel because we aren't going to be using it.

Updated the Blender Operation SaveObj in the Blender Operations method to actually use the filename provided by the user.