this is a fix for the development branch
There was an error in merging the branches 86 and 104. this fix removes the if checks in DefinePixels and is renamed to FindPixels.
FindPixels takes in the X and Y size of the image (ImageRow, ImageColumn)
The base is the value that either Row(if we are not vertical) or Column (column if we are vertical) are reset to.
the IteratorValue is how much the iterator is increased by every loop
ImageArea is amount of pixels in the image.
image is the image we are looping throuh
Color is the color we are trying to find in the image
Is Vertical determines if we are using the column(isVertical == True) to go through the list of the row(isVertical == False)