Added Adjacent Edge to Shape
I moved the transpose and the GetMidPoint function so they can be used inside of the meshCreating instead of after it. With these function inside the CalculateZAxis function we can mirror the mesh with the adjacent point.

Also I removed the extra images being generated when the program is ran