Contains the UI update
Changes the work flow of the program
The Mesh structure will no longer appear at the beginning , but a plane has to be confirmed first.

The user now no longer has to click twice to select an image

The confirm button will no longer appear without a plane being added.

Created Reset function inside of blender operations. This just resets the inputted images.

Moved MeshSettings to the ui_operations so we can have interations between the Views tab and the meshSettigns tabs

Changed the Views tab to the Image input Section.

Disabled the testing section